chore(exceptions): Moved custom exceptions into there own file
