+++ lib/exceptions/exceptions.dart
@@ -1,0 +1,15 @@
+class PermissionDeniedException implements Exception {
+  final String message;
+  PermissionDeniedException(this.message);
+
+  @override
+  String toString() => 'PermissionDeniedException: $message';
+}
+
+class ContactNotFoundException implements Exception {
+  final String message;
+  ContactNotFoundException(this.message);
+
+  @override
+  String toString() => 'ContactNotFoundException: $message';
+}
